+++ lib/async.js
@@ -1,9 +1,9 @@
 var checkParameters = require('./precondition')
 var defaultEncoding = require('./default-encoding')
 var sync = require('./sync')
-function zeroFill(num) {
+function zeroFill (num) {
   if (Buffer.alloc) {
-    return Buffer.alloc(num);
+    return Buffer.alloc(num)
   }
   var out = new Buffer(num)
   out.fill(0)
@@ -34,7 +34,7 @@ function checkNative (algo) {
     return checks[algo]
   }
   ZERO_BUF = ZERO_BUF || zeroFill(8)
-  let prom = browserPbkdf2(ZERO_BUF, ZERO_BUF, 10, 128, algo)
+  var prom = browserPbkdf2(ZERO_BUF, ZERO_BUF, 10, 128, algo)
     .then(function () {
       return true
     }).catch(function () {
