+++ lib/default-encoding.js
