+++ lib/precondition.js
